try doing some inner class bullshit
