+++ src/pieces/Bishop.java
@@ -41,7 +41,7 @@ public class Bishop extends GamePiece
     }
 
     @Override
-    public void updateValidMoves(Board board, ArrayList<GamePiece> moves)
+    public void updateValidMoves(Board board, ArrayList<Board.Move> moves)
     {
         final String COLOR = super.getColor();
         final int COL = super.getCol();
@@ -57,7 +57,7 @@ public class Bishop extends GamePiece
         {
             int spaceStatus = board.checkSpace(x, y, COLOR);
             if (spaceStatus == 1) { break; }
-            moves.add(new GamePiece("x", x, y));
+            moves.add(board.new Move(this, x, y));
             if (spaceStatus == 2) { break; }
             x--; y--;
         }
@@ -68,7 +68,7 @@ public class Bishop extends GamePiece
         {
             int spaceStatus = board.checkSpace(x, y, COLOR);
             if (spaceStatus == 1) { break; }
-            moves.add(new GamePiece("x", x, y));
+            moves.add(board.new Move(this, x, y));
             if (spaceStatus == 2) { break; }
             x++; y++;
         }
@@ -79,7 +79,7 @@ public class Bishop extends GamePiece
         {
             int spaceStatus = board.checkSpace(x, y, COLOR);
             if (spaceStatus == 1) { break; }
-            moves.add(new GamePiece("x", x, y));
+            moves.add(board.new Move(this, x, y));
             if (spaceStatus == 2) { break; }
             x--;
             y++;
@@ -91,7 +91,7 @@ public class Bishop extends GamePiece
         {
             int spaceStatus = board.checkSpace(x, y, COLOR);
             if (spaceStatus == 1) { break; }
-            moves.add(new GamePiece("x", x, y));
+            moves.add(board.new Move(this, x, y));
             if (spaceStatus == 2) { break; }
             x++; y--;
         }
